+++ shared/internal/port/v0.6.0/src/ports/data.ts
@@ -46,4 +46,7 @@ export default [
         customs: { dependsOn(p) { return p.dataSource === 'value' } }
     },
     { name: 'data', displayName: 'Data', group: 'Data', type: 'object' },
+    // For tableSelectionScope    
+    { name: 'selectionScope', displayName: 'selectionScope', group: 'Data', type: 'object' }, // MD
+    { name: 'selectionByTableId', displayName: 'selectionByTableId', group: 'Data', type: 'object' }, // MD
 ] as const satisfies readonly NodePort[]