+++ shared/internal/scope/v0.2.0/types.d.ts
@@ -1,0 +1,8 @@
+import { ColorScheme } from '@mantine/core'
+import { BaseReactProps } from '@shared/node'
+import { RolderType } from '@shared/types'
+
+export type CompProps = BaseReactProps & {
+    appLoaderColor?: string
+    colorScheme: ColorScheme | 'auto'    
+}