+++ shared/internal/types/v0.1.0/types.d.ts
@@ -105,7 +105,8 @@ export type Item = {
         createdAt: number
         updater: string | null
         updatedAt: number | null
-    }
+    },
+    parentTableId: string | undefined | null // MD
 }
 
 export type User = {