+++ CHANGELOG.md
@@ -1,5 +1,61 @@
 # Changelog
 
+## 2024-04-05 v1.0.0-beta21-table-selection-scope-1.0.0 by Vezdexod
+
+### mantine
+
+#### Table v1.2.0
+
+* Теперь при активном expension создается jotai атом и состояния multiSelection обрабатываются специальной функцией. Если обернуть главную таблицу (по сути все дерево таблиц) в tableSelectionScope, то можно будет получать и сбрасывать значение атома.
+* Все статусы обрабатываются с учетом иерархии
+* Изменены названия используемых scope в соответствии scope-0.2.0
+    * Сменились названия компонетов при создании scope 0.2.0
+        * cellScope на tableCellScope (Нода tableCell v1.1.0)
+        * tableScope на tableFilterScope (Нода tableFilter v1.1.0)
+        * Добавился tableCetableSelectionScopellScope
+* Выдает на выход свой id и id родительской таблицы (tableId, parentTableId).
+
+#### TableSelectionScope
+
+* Добавлена нода TableSelectionScope, с помощью которой можно получать значения id и статуса multiSelect всех записей, всех таблиц, которые расположены пол данной нодой (в иерархии). 
+* Можно сбрасывать селекты со всех таблиц иерархии.
+* Выдает на выход объект selectionByTableId, где ключи - tableId, а значения - массив выбранных запсией
+
+### Rolder Kit
+
+Созданы директории:
+rolder-kit/shared/internal/scope/v0.2.0
+rolder-kit/shared/components/ui/mantine/organisms/table/v1.2.0
+rolder-kit/shared/components/ui/mantine/organisms/table/modules/columnFilter/v1.1.0
+rolder-kit/shared/components/ui/mantine/organisms/table/modules/expansionRow/v1.1.0
+rolder-kit/shared/components/ui/mantine/molecules/tableScope/v1.0.0
+rolder-kit/nodes/ui/mantine/molecules/tableSelectionScope
+
+Изменены общие файлы: // MD
+* nodes/ui/mantine/mantine.ts
+  * ссылка на код: import tableSelectionScopeNode from "@nodes/table-selection-scope"
+  * и название ноды: tableSelectionScopeNode
+  * в rolder-kit/nodes/ui/mantine/package.json добавлена запись "@nodes/table-selection-scope": "workspace:*"
+
+* shared/internal/port/v0.6.0/src/ports/data.ts
+  * Содержимое атома tableSelectionScope; { name: 'selectionScope', displayName: 'selectionScope', group: 'Data', type: 'object' }
+  * Объект, где ключи - tableId, а значения - массив выбранных запсией: { name: 'selectionByTableId', displayName: 'selectionByTableId', group: 'Data', type: 'object' }
+* shared/internal/port/v0.6.0/src/ports/table.ts
+  * id текущей таблицы: { name: 'tableId', group: 'Data', type: 'string', displayName: 'tableId' }
+  * id родительской таблицы: { name: 'parentTableId', group: 'Data', type: 'string', displayName: 'parentTableId' }
+
+* shared/internal/types/v0.1.0/types.d.ts
+  * К типу Item добавлено поле для передачи tableId от родительсмкой таблицы дочерним: parentTableId: string | undefined | null
+* rolder-kit/nodes/ui/mantine/organisms/table/package.jsonrolder-kit/nodes/ui/mantine/organisms/table/package.json
+  * Добавлена ссылка на код с table-1.2.0: "@shared/table-v1.2.0": "workspace:*"
+  * Добавлена соответствующая версия в tableNode.ts
+* rolder-kit/nodes/ui/mantine/organisms/table/modules/expansionRow/package.json
+  * Добавлена ссылка на код с column-filter-v1.1.0: "@shared/column-filter-v1.1.0": "workspace:*"
+  * Добавлена соответствующая версия в columnFilterNode.ts
+* rolder-kit/nodes/ui/mantine/organisms/table/modules/expansionRow/package.json
+  * Добавлена ссылка на код с expansion-row-v1.1.0: "@shared/expansion-row-v1.1.0": "workspace:*"
+  * Добавлена соответствующая версия в expansionRowNode.ts
+
 ## 2024-04-01 v1.0.0-beta21
 
 ### mantine